Update to razzle 4
