+++ docs/source/developer-guidelines/typescript.md
@@ -70,7 +70,7 @@ Change your scripts in your ``package.json`` like this (inspired by [https://git
   "prettier": "./node_modules/.bin/prettier --single-quote --check 'src/**/*.{js,jsx,ts,tsx,json,css,scss,md}'",
   "prettier:fix": "./node_modules/.bin/prettier --single-quote --write 'src/**/*.{js,jsx,ts,tsx,json,css,scss,md}'",
   "prettier:ci": "./node_modules/.bin/prettier --single-quote --check 'src/**/*.{js,jsx,ts,tsx,json,css,scss,md}'",
-  "test": "razzle test --env=jest-environment-jsdom-sixteen --passWithNoTests",
+  "test": "razzle test --passWithNoTests",
   "start:prod": "NODE_ENV=production node build/server.js",
   "i18n": "rm -rf build/messages && NODE_ENV=production i18n",
   "develop": "missdev --output=addons --fetch-https"
