+++ package-why.json
@@ -3,9 +3,7 @@
     "postinstall": "Description for `npm run postinstall` command",
     "analyze": "Description for `npm run analyze` command",
     "start": "Description for `npm run start` command",
-    "start-spa": "Description for `npm run start-spa` command",
     "build": "Description for `npm run build` command",
-    "build-spa": "Description for `npm run build-spa` command",
     "build:clean": "Description for `npm run build:clean` command",
     "dist:clean": "Description for `npm run dist:clean` command",
     "build:dist": "Description for `npm run build:dist` command",
