+++ package.json
@@ -31,11 +31,9 @@
     "postinstall": "yarn patches",
     "analyze": "BUNDLE_ANALYZE=true razzle build",
     "start": "razzle start",
-    "start-spa": "razzle start --type=spa",
     "build": "razzle build",
-    "build-spa": "razzle build --type=spa",
-    "test": "razzle test --env=jest-environment-jsdom-sixteen --maxWorkers=50%",
-    "test:ci": "CI=true NODE_ICU_DATA=node_modules/full-icu razzle test --env=jest-environment-jsdom-sixteen",
+    "test": "razzle test --maxWorkers=50%",
+    "test:ci": "CI=true NODE_ICU_DATA=node_modules/full-icu razzle test",
     "test:husky": "CI=true yarn test --bail --findRelatedTests",
     "test:debug": "node --inspect node_modules/.bin/jest --runInBand",
     "start:prod": "NODE_ENV=production node build/server.js",
@@ -73,8 +71,6 @@
     ],
     "transform": {
       "^.+\\.js(x)?$": "babel-jest",
-      "^.+\\.css$": "jest-css-modules",
-      "^.+\\.scss$": "jest-css-modules",
       "^.+\\.(png)$": "jest-file",
       "^.+\\.(jpg)$": "jest-file",
       "^.+\\.(svg)$": "./jest-svgsystem-transform.js"
@@ -224,11 +220,12 @@
     "node": "^14 || ^16"
   },
   "dependencies": {
-    "@babel/plugin-proposal-export-default-from": "7.10.4",
-    "@babel/plugin-proposal-export-namespace-from": "7.10.4",
-    "@babel/plugin-proposal-json-strings": "7.10.4",
+    "@babel/core": "^7.0.0",
+    "@babel/plugin-proposal-export-default-from": "7.18.9",
+    "@babel/plugin-proposal-export-namespace-from": "7.18.9",
+    "@babel/plugin-proposal-json-strings": "7.18.6",
     "@babel/plugin-proposal-nullish-coalescing-operator": "7.18.6",
-    "@babel/plugin-proposal-throw-expressions": "7.10.4",
+    "@babel/plugin-proposal-throw-expressions": "7.18.6",
     "@loadable/babel-plugin": "5.13.2",
     "@loadable/component": "5.14.1",
     "@loadable/server": "5.14.0",
@@ -243,18 +240,17 @@
     "@testing-library/react": "12.1.5",
     "autoprefixer": "9.7.6",
     "axe-core": "4.4.2",
-    "babel-7-plugin-root-import": "5.4.0",
     "babel-eslint": "10.1.0",
     "babel-plugin-add-module-exports": "0.2.1",
     "babel-plugin-lodash": "3.3.4",
     "babel-plugin-react-intl": "5.1.17",
     "babel-plugin-root-import": "6.1.0",
+    "babel-preset-razzle": "4.2.17",
     "bundlewatch": "0.2.7",
     "circular-dependency-plugin": "5.2.2",
     "classnames": "2.2.6",
     "commander": "8.2.0",
     "connected-react-router": "6.8.0",
-    "coveralls": "3.0.2",
     "crypto-random-string": "3.2.0",
     "cypress": "10.3.1",
     "cypress-axe": "0.14.0",
@@ -287,17 +283,13 @@
     "express": "4.17.1",
     "glob": "7.1.6",
     "hamburgers": "1.1.3",
-    "handlebars": "4.7.7",
     "history": "4.10.1",
     "http-proxy-middleware": "2.0.1",
     "husky": "4.2.5",
     "identity-obj-proxy": "3.0.0",
     "image-extensions": "1.1.0",
     "is-url": "1.2.4",
-    "jest-css-modules": "1.1.0",
-    "jest-environment-jsdom-sixteen": "1.0.3",
     "jest-file": "1.0.0",
-    "js-yaml": "3.13.1",
     "jsonwebtoken": "8.3.0",
     "jwt-decode": "2.2.0",
     "less": "3.11.1",
@@ -308,8 +300,6 @@
     "lodash": "4.17.21",
     "lodash-move": "1.1.1",
     "lodash-webpack-plugin": "0.11.5",
-    "luxon": "1.22.0",
-    "mem": "4.0.0",
     "mini-css-extract-plugin": "0.9.0",
     "moment": "2.24.0",
     "pofile": "1.0.10",
@@ -324,8 +314,9 @@
     "promise-file-reader": "1.0.2",
     "prop-types": "15.7.2",
     "query-string": "7.1.0",
-    "razzle": "3.4.5",
-    "razzle-plugin-bundle-analyzer": "1.2.0",
+    "razzle": "4.2.17",
+    "razzle-dev-utils": "4.2.17",
+    "razzle-plugin-bundle-analyzer": "4.2.17",
     "rc-time-picker": "3.7.3",
     "react": "17.0.2",
     "react-anchor-link-smooth-scroll": "1.0.12",
@@ -375,7 +366,6 @@
     "slate-history": "0.66.0",
     "slate-hyperscript": "0.77.0",
     "slate-react": "0.78.1",
-    "start-server-and-test": "1.14.0",
     "stylelint": "14.0.1",
     "stylelint-config-idiomatic-order": "8.1.0",
     "stylelint-config-prettier": "8.0.1",
@@ -398,15 +388,18 @@
     "@storybook/addon-essentials": "^6.3.0",
     "@storybook/addon-links": "^6.3.0",
     "@storybook/react": "^6.3.0",
-    "babel-loader": "^8.1.0",
+    "babel-loader": "8.2.5",
     "full-icu": "1.4.0",
+    "html-webpack-plugin": "4.5.2",
     "identity-obj-proxy": "3.0.0",
-    "jest": "27.4.5",
-    "jest-environment-jsdom-sixteen": "1.0.3",
+    "jest": "26.6.3",
     "react-is": "^16.13.1",
     "release-it": "^15.1.3",
     "tmp": "0.2.1",
     "use-trace-update": "1.3.2",
+    "webpack": "4.46.0",
+    "webpack-dev-server": "3.11.0",
+    "webpack-node-externals": "3.0.0",
     "why": "0.6.2"
   },
   "resolutions-comments": {
