+++ razzle.config.js
@@ -5,7 +5,6 @@ const nodeExternals = require('webpack-node-externals');
 const LoadablePlugin = require('@loadable/webpack-plugin');
 const LodashModuleReplacementPlugin = require('lodash-webpack-plugin');
 const fs = require('fs');
-const { pickBy } = require('lodash');
 const RootResolverPlugin = require('./webpack-plugins/webpack-root-resolver');
 const RelativeResolverPlugin = require('./webpack-plugins/webpack-relative-resolver');
 const createAddonsLoader = require('./create-addons-loader');
@@ -13,11 +12,10 @@ const AddonConfigurationRegistry = require('./addon-registry');
 const CircularDependencyPlugin = require('circular-dependency-plugin');
 
 const fileLoaderFinder = makeLoaderFinder('file-loader');
-const babelLoaderFinder = makeLoaderFinder('babel-loader');
 
 const projectRootPath = path.resolve('.');
 const languages = require('./src/constants/Languages');
-const { poToJson } = require('@plone/scripts/i18n');
+const { poToJson } = require('./packages/scripts/i18n'); // FIXME
 
 const packageJson = require(path.join(projectRootPath, 'package.json'));
 
@@ -27,6 +25,7 @@ const defaultModify = ({
   env: { target, dev },
   webpackConfig: config,
   webpackObject: webpack,
+  options,
 }) => {
   // Compile language JSON files from po files
   poToJson({ registry, addonMode: false });
@@ -188,9 +187,6 @@ const defaultModify = ({
     ...fileLoader.exclude,
   ];
 
-  // Disabling the ESlint pre loader
-  config.module.rules.splice(0, 1);
-
   let addonsFromEnvVar = [];
   if (process.env.ADDONS) {
     addonsFromEnvVar = process.env.ADDONS.split(';');
@@ -235,8 +231,7 @@ const defaultModify = ({
 
   let addonsAsExternals = [];
 
-  const babelLoader = config.module.rules.find(babelLoaderFinder);
-  const { include } = babelLoader;
+  const { include } = options.webpackOptions.babelRule;
   if (packageJson.name !== '@plone/volto') {
     include.push(fs.realpathSync(`${registry.voltoPath}/src`));
   }
@@ -275,13 +270,13 @@ const defaultModify = ({
     target === 'node'
       ? [
           nodeExternals({
-            whitelist: [
+            allowlist: [
               dev ? 'webpack/hot/poll?300' : null,
               /\.(eot|woff|woff2|ttf|otf)$/,
               /\.(svg|png|jpg|jpeg|gif|ico)$/,
               /\.(mp4|mp3|ogg|swf|webp)$/,
               /\.(css|scss|sass|sss|less)$/,
-              // Add support for whitelist external (ie. node_modules npm published packages)
+              // Add support for addons to include externals (ie. node_modules npm published packages)
               ...addonsAsExternals,
               /^@plone\/volto/,
             ].filter(Boolean),
@@ -309,15 +304,21 @@ const plugins = addonExtenders.reduce(
 
 module.exports = {
   plugins,
+  modifyJestConfig: ({ jestConfig }) => {
+    jestConfig.testEnvironment = 'jsdom';
+    return jestConfig;
+  },
   modifyWebpackConfig: ({
     env: { target, dev },
     webpackConfig,
     webpackObject,
+    options,
   }) => {
     const defaultConfig = defaultModify({
       env: { target, dev },
       webpackConfig,
       webpackObject,
+      options,
     });
 
     const res = addonExtenders.reduce(
@@ -326,7 +327,7 @@ module.exports = {
     );
     return res;
   },
-  experimental: {
-    reactRefresh: true,
+  options: {
+    enableReactRefresh: true,
   },
 };
